# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon2026 Backend API
  version: 1.3.0
  description: |
    OpenAPI contract for the Hackathon2026 backend.

    Authentication model:
    - `POST /auth/psychologists/login` sets `session_id` and `csrf_token` cookies.
    - `POST /auth/admin/login` sets `admin_session_id` and `csrf_token` cookies.
    - Protected `GET` endpoints require the corresponding session cookie.
    - Protected non-`GET` endpoints also require header `X-CSRF-Token` with the same value as the `csrf_token` cookie.

    Practical notes for frontend integration:
    - Base URL in local development is `http://localhost:8080`.
    - Error responses use the unified shape `{ "message": "...", "field_errors": { ... } }`.
    - Browser CORS preflight is supported for configured `ALLOWED_ORIGINS`, including credentials and the `X-CSRF-Token` header.
    - Psychologist user payloads return `portal_access_until` and `blocked_until` as `null` when they are not set.
    - Psychologist user payloads may include computed `account_status` and `subscription_status`.
    - `GET /psychologists/tests` returns activity metrics such as started/completed counts and last activity timestamps for the “Мои опросники” dashboard.
    - Test creation defaults `status` to `published` when omitted.
    - Test update defaults `status` to `draft` when omitted.
    - Public test start requires `respondent_name` and `respondent_phone`.
...
    - In-progress public sessions expire after 1 hour of inactivity.
...
servers:
...
paths:
...
  /auth/admin/login:
    post:
      tags:
        - Admin Auth
      summary: Login admin
      description: Creates an admin session and issues both session and CSRF cookies.
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginInput'
      responses:
        '200':
          description: Authenticated admin session created
          headers:
            Set-Cookie:
              description: "Sets `admin_session_id` and `csrf_token` cookies."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/admin/logout:
    post:
      tags:
        - Admin Auth
      summary: Logout admin
      operationId: logoutAdmin
      security:
        - adminSession: []
          csrfToken: []
      responses:
        '200':
          description: Session cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admins/me:
    get:
      tags:
        - Admin
      summary: Get current admin
      operationId: getAdminMe
      security:
        - adminSession: []
      responses:
        '200':
          description: Current admin account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Admin
      summary: Update current admin
      operationId: updateAdminMe
      security:
        - adminSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminMeInput'
      responses:
        '200':
          description: Updated admin account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    psychologistSession:
      type: apiKey
      in: cookie
      name: session_id
      description: Psychologist session cookie returned by the psychologist login endpoint.
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session_id
      description: Admin session cookie returned by the admin login endpoint.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: "Must match the `csrf_token` cookie value on protected non-GET requests."
...
  responses:
    BadRequest:
      description: Invalid request payload or path parameter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired session credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated request is not allowed or fails CSRF/origin checks.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
        - field_errors
      properties:
        message:
          type: string
          example: Validation failed
        field_errors:
          type: object
          nullable: true
          additionalProperties:
            type: string
          example:
            email: Email already exists
    StatusResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    User:
      type: object
      required:
        - id
        - email
        - full_name
        - role
        - is_active
        - portal_access_until
        - blocked_until
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
          description: Present for admin accounts, empty for psychologists.
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum:
            - admin
            - psychologist
            - client
        is_active:
          type: boolean
        portal_access_until:
          type: string
          format: date-time
          nullable: true
          description: "`null` when not set."
        blocked_until:
          type: string
          format: date-time
          nullable: true
          description: "`null` when not set."
        account_status:
          type: string
          description: Computed status used by the frontend for psychologist accounts.
          enum:
            - active
            - blocked
        subscription_status:
          type: string
          description: Computed subscription status used by the frontend for psychologist accounts.
          enum:
            - active
            - blocked
            - expired
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AdminLoginInput:
      type: object
      additionalProperties: false
      required:
        - login
        - password
      properties:
        login:
          type: string
          minLength: 1
        password:
          type: string
    AdminAuthResponse:
      type: object
      required:
        - expires_at
        - user
      properties:
        expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
...
    UpdateAdminMeInput:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email